requirements.txt
Flask==1.0.2
line-bot-sdk==1.8.0
gunicorn==19.9.0
gensim==3.2.0
jieba==0.39
sklearn==0.0
scipy==1.0.0